create "simulation" subpackage rename BMCTool class to BMCSim move BlochMcConnellSolver, BMCSim, simulate into simulation adjust imports
