+++ src/bmctool/simulation/_BlochMcConnellSolver.py
