+++ src/bmctool/simulation/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ['BlochMcConnellSolver', 'BMCSim', 'simulate']
+
+from bmctool.simulation._BlochMcConnellSolver import BlochMcConnellSolver
+from bmctool.simulation._BMCSim import BMCSim
+from bmctool.simulation._simulate import simulate
